Keep code compatible with Python <3.10
The `slots` parameter was added to dataclasses in Python 3.10. This commit changes the code to explicitly set the `__slots__` attribute instead of realying on `@dataclass(slots=True)`

The benchmark results are not affected by the change. The `test_structs_slots_dataclass_access` benchmark still reports a mean time of 43.8826ns